requirements.txt
fastapi==0.52.0
uvicorn==0.27.0
firebase-admin==6.5.0
python-dotenv==1.0.1
google==3.0.0
pydantic==1.10.14
google-auth-oauthlib==1.2.0 
oauthlib==3.2.2 
requests-oauthlib==2.0.0
jose==1.0.0